refactor: standardize version types to Long across codebase
Change version-related parameters and return types from Int to Long
for consistency with ProductEvent.version (Long). This eliminates
type casting at the domain layer while maintaining compatibility
with the database INTEGER columns through controlled conversions
at the repository layer.

Changes:
- ExtendedEventStoreRepository: fromVersion and getStreamVersion use Long
- EventStoreVersionConflictException: expectedVersion/actualVersion are Long
- ProductEventStoreRepository: internal version handling uses Long
- Integration tests: updated to use Long literals (1L, 2L, etc.)

+++ cqrs-spike/src/main/kotlin/com/pintailconsultingllc/cqrsspike/infrastructure/eventstore/EventStoreExceptions.kt
@@ -24,8 +24,8 @@ class EventStreamNotFoundException(
 class EventStoreVersionConflictException(
     val aggregateType: String,
     val aggregateId: UUID,
-    val expectedVersion: Int,
-    val actualVersion: Int
+    val expectedVersion: Long,
+    val actualVersion: Long
 ) : EventStoreException(
     "Version conflict for $aggregateType/$aggregateId: expected $expectedVersion, actual $actualVersion"
 )
